+++ BLL/Common/Geometry/Point2D.cs
@@ -11,19 +11,19 @@ namespace RDBLL.Common.Geometry
 {
     public class Point2D : ICloneable
     {
-        public decimal Coord_X { get; set; }
-        public decimal Coord_Y { get; set; }
+        public double Coord_X { get; set; }
+        public double Coord_Y { get; set; }
 
         public String PointText()
         {
             return "(" + Convert.ToString(Coord_X) + ";" + Convert.ToString(Coord_Y) + ")";
         }
 
-        public Point2D EndPoint (decimal Angle, decimal Length)
+        public Point2D EndPoint (double Angle, double Length)
         {
             Point2D EndPoint = new Point2D(0,0);
-            EndPoint.Coord_X = this.Coord_X + Convert.ToDecimal(Math.Cos(Convert.ToDouble(Angle))) * Length;
-            EndPoint.Coord_Y = this.Coord_Y + Convert.ToDecimal(Math.Sin(Convert.ToDouble(Angle))) * Length;
+            EndPoint.Coord_X = this.Coord_X + Convert.ToDouble(Math.Cos(Convert.ToDouble(Angle))) * Length;
+            EndPoint.Coord_Y = this.Coord_Y + Convert.ToDouble(Math.Sin(Convert.ToDouble(Angle))) * Length;
             return EndPoint;
         }
 
@@ -35,7 +35,7 @@ namespace RDBLL.Common.Geometry
             return xmlNode;
         }
 
-        public Point2D(decimal coord_X, decimal coord_Y)
+        public Point2D(double coord_X, double coord_Y)
         {
             Coord_X = coord_X;
             Coord_Y = coord_Y;
@@ -45,8 +45,8 @@ namespace RDBLL.Common.Geometry
         {
             foreach (XmlAttribute obj in xmlNode.Attributes)
             {
-                if (obj.Name == "Coord_X") Coord_X = Convert.ToDecimal(obj.Value);
-                if (obj.Name == "Coord_Y") Coord_Y = Convert.ToDecimal(obj.Value);
+                if (obj.Name == "Coord_X") Coord_X = Convert.ToDouble(obj.Value);
+                if (obj.Name == "Coord_Y") Coord_Y = Convert.ToDouble(obj.Value);
             }
         }
 
@@ -61,7 +61,7 @@ namespace RDBLL.Common.Geometry
         public Point2D StartPoint { get; set; }
         public Point2D EndPoint { get; set; }
 
-        public decimal GetDistance(Point2D StartPoint, Point2D EndPoint)
+        public double GetDistance(Point2D StartPoint, Point2D EndPoint)
         {
             return (StartPoint.Coord_X - EndPoint.Coord_X);
                 //Math.Sqrt(Convert.ToDouble(StartPoint.coord_X - EndPoint.coord_X));
@@ -70,11 +70,11 @@ namespace RDBLL.Common.Geometry
 
     public class Geometry2D
     {
-        public decimal GetDistance(Point2D StartPoint, Point2D EndPoint)
+        public double GetDistance(Point2D StartPoint, Point2D EndPoint)
         {
-            decimal dX = EndPoint.Coord_X - StartPoint.Coord_X;
-            decimal dY = EndPoint.Coord_Y - StartPoint.Coord_Y;
-            return Convert.ToDecimal(Math.Sqrt(Convert.ToDouble(dX*dX + dY*dY)));
+            double dX = EndPoint.Coord_X - StartPoint.Coord_X;
+            double dY = EndPoint.Coord_Y - StartPoint.Coord_Y;
+            return Convert.ToDouble(Math.Sqrt(Convert.ToDouble(dX*dX + dY*dY)));
         }
     }
 }
